Solved conflicts:
	bitbake/lib/bb/command.py
	bitbake/lib/bb/fetch2/__init__.py
	bitbake/lib/bb/fetch2/perforce.py
	bitbake/lib/bb/monitordisk.py
	documentation/adt-manual/adt-command.xml
	documentation/adt-manual/adt-intro.xml
	documentation/adt-manual/adt-manual.xml
	documentation/adt-manual/adt-prepare.xml
	documentation/bsp-guide/bsp-guide.xml
	documentation/dev-manual/dev-manual-common-tasks.xml
	documentation/dev-manual/dev-manual-model.xml
	documentation/dev-manual/dev-manual-newbie.xml
	documentation/dev-manual/dev-manual-start.xml
	documentation/dev-manual/dev-manual.xml
	documentation/kernel-dev/kernel-dev.xml
	documentation/poky.ent
	documentation/profile-manual/profile-manual.xml
	documentation/ref-manual/closer-look.xml
	documentation/ref-manual/faq.xml
	documentation/ref-manual/introduction.xml
	documentation/ref-manual/migration.xml
	documentation/ref-manual/ref-classes.xml
	documentation/ref-manual/ref-features.xml
	documentation/ref-manual/ref-images.xml
	documentation/ref-manual/ref-manual.xml
	documentation/ref-manual/ref-structure.xml
	documentation/ref-manual/ref-variables.xml
	documentation/ref-manual/technical-details.xml
	documentation/ref-manual/usingpoky.xml
	documentation/tools/mega-manual.sed
	documentation/yocto-project-qs/yocto-project-qs.xml
	meta-yocto-bsp/recipes-kernel/linux/linux-yocto_3.10.bbappend
	meta-yocto/conf/distro/include/package_regex.inc
	meta-yocto/conf/distro/poky.conf
	meta/classes/base.bbclass
	meta/classes/boot-directdisk.bbclass
	meta/classes/bootimg.bbclass
	meta/classes/cross-canadian.bbclass
	meta/classes/grub-efi.bbclass
	meta/classes/gummiboot.bbclass
	meta/classes/kernel-yocto.bbclass
	meta/classes/kernel.bbclass
	meta/classes/license.bbclass
	meta/classes/package_rpm.bbclass
	meta/classes/ptest.bbclass
	meta/classes/rootfs_deb.bbclass
	meta/classes/rootfs_ipk.bbclass
	meta/classes/rootfs_rpm.bbclass
	meta/files/device_table-minimal.txt
	meta/recipes-bsp/gnu-efi/gnu-efi_3.0u.bb
	meta/recipes-bsp/grub/grub-efi-native_2.00.bb
	meta/recipes-bsp/grub/grub_2.00.bb
	meta/recipes-bsp/gummiboot/gummiboot_git.bb
	meta/recipes-connectivity/nfs-utils/nfs-utils/nfsserver
	meta/recipes-connectivity/nfs-utils/nfs-utils_1.2.8.bb
	meta/recipes-connectivity/openssh/openssh-6.2p2/init
	meta/recipes-connectivity/openssl/openssl.inc
	meta/recipes-connectivity/openssl/openssl_1.0.1e.bb
	meta/recipes-connectivity/ppp/ppp_2.4.5.bb
	meta/recipes-connectivity/wpa-supplicant/wpa-supplicant/defconfig-gnutls
	meta/recipes-core/busybox/busybox-1.21.1/defconfig
	meta/recipes-core/busybox/busybox_1.21.1.bb
	meta/recipes-core/dbus/dbus-1.6.10/dbus-1.init
	meta/recipes-core/eglibc/eglibc_2.17.bb
	meta/recipes-core/eglibc/eglibc_2.18.bb
	meta/recipes-core/images/build-appliance-image_8.0.bb
	meta/recipes-core/initrdscripts/files/init-install-efi.sh
	meta/recipes-core/initscripts/initscripts_1.0.bb
	meta/recipes-core/meta/buildtools-tarball.bb
	meta/recipes-core/packagegroups/packagegroup-self-hosted.bb
	meta/recipes-core/systemd/systemd-compat-units.bb
	meta/recipes-core/systemd/systemd-compat-units/run-postinsts.service
	meta/recipes-core/sysvinit/sysvinit-inittab_2.88dsf.bb
	meta/recipes-core/udev/udev.inc
	meta/recipes-devtools/binutils/binutils.inc
	meta/recipes-devtools/e2fsprogs/e2fsprogs-1.42.8/populate-extfs.sh
	meta/recipes-devtools/flex/flex.inc
	meta/recipes-devtools/gdb/gdb-7.6.inc
	meta/recipes-devtools/mklibs/mklibs-native_0.1.38.bb
	meta/recipes-devtools/opkg/opkg.inc
	meta/recipes-devtools/perl/perl-ptest.inc
	meta/recipes-devtools/pseudo/pseudo.inc
	meta/recipes-devtools/python/python-2.7-manifest.inc
	meta/recipes-devtools/rpm/rpm/debugedit-valid-file-to-fix-segment-fault.patch
	meta/recipes-devtools/rpm/rpm_5.4.9.bb
	meta/recipes-extended/cracklib/cracklib_2.9.0.bb
	meta/recipes-extended/ethtool/ethtool_3.10.bb
	meta/recipes-extended/images/core-image-basic.bb
	meta/recipes-extended/mdadm/mdadm_3.2.6.bb
	meta/recipes-extended/pigz/pigz_2.3.bb
	meta/recipes-graphics/cairo/cairo.inc
	meta/recipes-graphics/pango/pango.inc
	meta/recipes-graphics/xorg-lib/xorg-lib-common.inc
	meta/recipes-kernel/kexec/kexec-tools_2.0.7.bb
	meta/recipes-kernel/kmod/kmod-native_git.bb
	meta/recipes-kernel/kmod/kmod.inc
	meta/recipes-kernel/kmod/kmod/Change-to-calling-bswap_-instead-of-htobe-and-be-toh.patch
	meta/recipes-kernel/linux/linux-yocto-rt_3.10.bb
	meta/recipes-kernel/linux/linux-yocto-rt_3.8.bb
	meta/recipes-kernel/linux/linux-yocto-tiny_3.10.bb
	meta/recipes-kernel/linux/linux-yocto-tiny_3.8.bb
	meta/recipes-kernel/linux/linux-yocto_3.10.bb
	meta/recipes-kernel/linux/linux-yocto_3.8.bb
	meta/recipes-kernel/lttng/lttng-modules_2.3.3.bb
	meta/recipes-multimedia/gstreamer/gst-ffmpeg-0.10.13/gst-ffmpeg-CVE-2013-3674.patch
	meta/recipes-multimedia/gstreamer/gst-ffmpeg_0.10.13.bb
	meta/recipes-multimedia/libtiff/tiff_4.0.3.bb
	meta/recipes-qt/meta/meta-toolchain-qt.inc
	meta/recipes-qt/qt4/qt4-4.8.5.inc
	meta/recipes-qt/qt4/qt4.inc
	meta/recipes-support/boost/boost_1.54.0.bb
	meta/recipes-support/libnl/libnl_3.2.22.bb
	scripts/contrib/python/generate-manifest-2.7.py
	scripts/lib/mic/conf.py
	scripts/lib/mic/creator.py
	scripts/lib/mic/imager/baseimager.py
	scripts/lib/mic/imager/livecd.py
	scripts/lib/mic/kickstart/custom_commands/partition.py
	scripts/lib/mic/pluginbase.py
	scripts/lib/mic/plugins/imager/direct_plugin.py
	scripts/lib/mic/plugins/imager/fs_plugin.py
	scripts/lib/mic/plugins/imager/livecd_plugin.py
	scripts/lib/mic/plugins/imager/liveusb_plugin.py
	scripts/lib/mic/plugins/imager/loop_plugin.py
	scripts/lib/mic/plugins/imager/raw_plugin.py
	scripts/lib/mic/utils/fs_related.py
	scripts/lib/mic/utils/misc.py
	scripts/lib/mic/utils/partitionedfs.py
	scripts/runqemu
	scripts/runqemu-internal
	scripts/wic

+++ bitbake/lib/prserv/serv.py
@@ -13,6 +13,7 @@ except ImportError:
 import bb.server.xmlrpc
 import prserv
 import prserv.db
+import errno
 
 logger = logging.getLogger("BitBake.PRserv")
 
@@ -37,8 +38,17 @@ singleton = None
 class PRServer(SimpleXMLRPCServer):
     def __init__(self, dbfile, logfile, interface, daemon=True):
         ''' constructor '''
-        SimpleXMLRPCServer.__init__(self, interface,
-                                    logRequests=False, allow_none=True)
+        import socket
+        try:
+            SimpleXMLRPCServer.__init__(self, interface,
+                                        logRequests=False, allow_none=True)
+        except socket.error:
+            ip=socket.gethostbyname(interface[0])
+            port=interface[1]
+            msg="PR Server unable to bind to %s:%s\n" % (ip, port)
+            sys.stderr.write(msg)
+            raise PRServiceConfigError
+
         self.dbfile=dbfile
         self.daemon=daemon
         self.logfile=logfile
@@ -66,11 +76,19 @@ class PRServer(SimpleXMLRPCServer):
         In addition, exception handling is done here.
 
         """
+        iter_count = 1
+        # With 60 iterations between syncs and a 0.5 second timeout between
+        # iterations, this will sync if dirty every ~30 seconds.
+        iterations_between_sync = 60
+
         while True:
             (request, client_address) = self.requestqueue.get()
             try:
                 self.finish_request(request, client_address)
                 self.shutdown_request(request)
+                iter_count = (iter_count + 1) % iterations_between_sync
+                if iter_count == 0:
+                    self.table.sync_if_dirty()
             except:
                 self.handle_error(request, client_address)
                 self.shutdown_request(request)
@@ -274,20 +292,37 @@ def stop_daemon(host, port):
         PRServerConnection(host, port).terminate()
     except:
         logger.critical("Stop PRService %s:%d failed" % (host,port))
-    time.sleep(0.5)
 
     try:
         if pid:
+            wait_timeout = 0
+            print("Waiting for pr-server to exit.")
+            while is_running(pid) and wait_timeout < 50:
+                time.sleep(0.1)
+                wait_timeout += 1
+
+            if is_running(pid):
+                print("Sending SIGTERM to pr-server.")
+                os.kill(pid,signal.SIGTERM)
+                time.sleep(0.1)
+
             if os.path.exists(pidfile):
                 os.remove(pidfile)
-            os.kill(pid,signal.SIGTERM)
-            time.sleep(0.1)
+
     except OSError as e:
         err = str(e)
         if err.find("No such process") <= 0:
             raise e
 
     return 0
+
+def is_running(pid):
+    try:
+        os.kill(pid, 0)
+    except OSError as err:
+        if err.errno == errno.ESRCH:
+            return False
+    return True
 
 def is_local_special(host, port):
     if host.strip().upper() == 'localhost'.upper() and (not port):
